+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [artifact-manager] IAC-788 / Improved Handling of Empty vRA Blueprint Versions for vRA 8.12.x.
 
 ### Enhancements
+* [vro-types/o11n-plugin-aria] IAC-789 / Adding new Aria Automation Plugin vRO inventory Types
 * [artifact-manager] IAC-786 / Set Aria Operations Default Policy vROPs 8.12.x.
 * [artifact-manager] IAC-790 / Update usage of deprecated policy APIs for vROPs 8.12.x.
 * [archetypes] IAC-797 / Improved content.yaml definition for "ts-vra-ng" and "vra-ng"